Create SotM2020_sponsorship_brochure_text_draft1_3oct19.txt
I've made a start on the sponsorship brochure by just writing out the text in the Heidelberg 2019 version. Next I'd like to convert it to a LaTeX file and start wit the layout and design. Pretty much a noob at github so this is my test document. Would like to figure out how to create folders for things like logo images etc (so not the coding stuff and doesn't really work with merges for the most part, I think) as well, so please advise if this propose new file stuff is the right way to go about it.

Former-commit-id: adf771c714baa074c1c05afbab1ddcfee6155360